Put auth init stuff in a hook, new component for centering on journey and mapillary position.
