+++ src/components/App.js
@@ -16,12 +16,12 @@ import get from "lodash/get";
 import {withRoute} from "../hoc/withRoute";
 import Graph from "./map/Graph";
 import LoginModal from "./LoginModal";
-import {checkExistingSession, authorize} from "../auth/authService";
-import {removeAuthParams} from "../stores/UrlManager";
 import ServerMessage from "./ServerMessage";
 import MapEvents from "./map/MapEvents";
 import LeafletMap from "./map/Map";
 import StopsByRouteQuery from "../queries/StopsByRouteQuery";
+import {useAuth} from "../auth/useAuth";
+import CenterOnPosition from "./map/CenterOnPosition";
 
 const AppFrame = styled.main`
   width: 100%;
@@ -81,41 +81,10 @@ function App({route, state, UI}) {
     journeyDetailsOpen,
   } = state;
 
+  // Receive the auth code and check session
+  useAuth(UI.setUser);
+
   const selectedJourneyId = getJourneyId(selectedJourney);
-  const {code, is_test = "false"} = useMemo(
-    () =>
-      Array.from(new URL(window.location.href).searchParams.entries()).reduce(
-        (params, [key, value]) => {
-          params[key] = value;
-          return params;
-        },
-        {}
-      ),
-    []
-  );
-
-  useEffect(() => {
-    const auth = async () => {
-      const response = await checkExistingSession();
-      response && response.isOk && response.email
-        ? UI.setUser(response.email)
-        : UI.setUser(null);
-
-      if (code) {
-        const response = await authorize(code, is_test === "true");
-
-        if (response && response.isOk && response.email) {
-          UI.setUser(response.email);
-        } else {
-          console.error("Login not successful.");
-        }
-
-        removeAuthParams();
-      }
-    };
-
-    auth();
-  }, [code, is_test]);
 
   // Condition for when the side panel is actually open, not only when it could be open.
   const detailsAreOpen = useMemo(
@@ -152,6 +121,7 @@ function App({route, state, UI}) {
                   <JourneyPosition date={date} journeys={routeAndSelected}>
                     {(currentJourneyPositions) => (
                       <>
+                        <CenterOnPosition journeyPositions={currentJourneyPositions} />
                         <SidePanel
                           areaJourneysLoading={!live && areaJourneysLoading}
                           journeyLoading={selectedJourneyLoading}
@@ -164,62 +134,16 @@ function App({route, state, UI}) {
                         <Map detailsOpen={detailsAreOpen}>
                           <SingleStopQuery date={date} stopId={selectedStopId}>
                             {({stop}) => (
-                              <>
-                                <Observer>
-                                  {() => {
-                                    if (live) {
-                                      return null;
-                                    }
-
-                                    // Set the map center from a selected
-                                    // stop position or selected
-                                    // selectedJourney position.
-                                    const stopPosition = stop
-                                      ? latLng([stop.lat, stop.lng])
-                                      : false;
-
-                                    const selectedJourneyPosition =
-                                      currentJourneyPositions.size === 1 &&
-                                      selectedJourneyId
-                                        ? currentJourneyPositions.get(
-                                            selectedJourneyId
-                                          ) || false
-                                        : false;
-
-                                    const {lat, lng} = selectedJourneyPosition || {};
-
-                                    // If a journey is selected, use the
-                                    // journey position if available.
-                                    // Else use the selected stop
-                                    // position if available.
-                                    let centerPosition = false;
-
-                                    if (state.currentMapillaryMapLocation) {
-                                      centerPosition = state.currentMapillaryMapLocation;
-                                    } else if (lat && lng && selectedJourney) {
-                                      centerPosition = latLng([lat, lng]);
-                                    } else if (!selectedJourney) {
-                                      centerPosition = stopPosition;
-                                    }
-
-                                    if (centerPosition) {
-                                      UI.setMapView(centerPosition);
-                                    }
-
-                                    return null;
-                                  }}
-                                </Observer>
-                                <MapContent
-                                  centerOnRoute={areaJourneys.length === 0}
-                                  routeJourneys={routeJourneys}
-                                  journeys={currentJourneys}
-                                  journeyPositions={currentJourneyPositions}
-                                  unsignedEvents={unsignedEvents}
-                                  route={route}
-                                  stop={stop}
-                                  routeStops={routeStops}
-                                />
-                              </>
+                              <MapContent
+                                centerOnRoute={areaJourneys.length === 0}
+                                routeJourneys={routeJourneys}
+                                journeys={currentJourneys}
+                                journeyPositions={currentJourneyPositions}
+                                unsignedEvents={unsignedEvents}
+                                route={route}
+                                stop={stop}
+                                routeStops={routeStops}
+                              />
                             )}
                           </SingleStopQuery>
                           {selectedJourney && (
